Refactored eval'd expression to support many props
In this change, I refactored the stringified function passed to the
DevTools API to support many properties. This allows for a slightly
clearer presentation of what the user is looking at.

In addition, I have added a function that finds the view instance that
corresponds to the currently selected element, if any.